backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
sqlalchemy>=2.0.30
asyncpg>=0.29.0
alembic>=1.13.1
pydantic[email]>=2.7.1
pydantic-settings>=2.2.1
email-validator>=2.0.0
passlib[bcrypt]
bcrypt==3.2.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
boto3>=1.34.106
celery>=5.4.0
redis>=5.0.4
httpx>=0.27.0
emails>=0.6
Jinja2>=3.1.4
